refactor: extract scoring into nightshift/scoring.py per code structure rules
Follows CLAUDE.md "one concern per module" rule:
- scoring.py: score_diff(), _diff_line_score(), _has_test_files()
- constants.py: SECURITY_PATTERNS, ERROR_HANDLING_PATTERNS, CATEGORY_SCORES
- cycle.py: no longer contains scoring logic or hardcoded pattern data
- install.sh: scoring.py added to PACKAGE_FILES
- Tests updated to import from nightshift.scoring
