+++ nightshift/cli.py
@@ -19,10 +19,10 @@ from nightshift.cycle import (
     extract_json,
     parse_cycle_result,
     recent_hot_files,
-    score_diff,
     verify_cycle,
 )
 from nightshift.errors import NightshiftError
+from nightshift.scoring import score_diff
 from nightshift.shell import command_exists, git, run_command
 from nightshift.state import append_cycle_state, load_json, read_state, write_json
 from nightshift.worktree import (
