+++ nightshift/cycle.py
@@ -16,7 +16,6 @@ from nightshift.state import top_path
 from nightshift.types import (
     CycleResult,
     CycleVerification,
-    DiffScore,
     NightshiftConfig,
     ShiftState,
 )
@@ -365,121 +364,3 @@ def verify_cycle(
             violations.append(f"Verification command failed: `{verify_command}`")
 
     return (not violations), verification
-
-
-# --- Diff Scorer -------------------------------------------------------------
-
-# Keywords in diff lines that signal high-impact changes, mapped to base scores.
-_SECURITY_PATTERNS: list[tuple[re.Pattern[str], int]] = [
-    (re.compile(r"(sql.?inject|xss|csrf|auth|sanitiz|escap|vulnerab)", re.IGNORECASE), 8),
-    (re.compile(r"(password|secret|token|credential|api.?key)", re.IGNORECASE), 7),
-]
-
-_ERROR_HANDLING_PATTERNS: list[tuple[re.Pattern[str], int]] = [
-    (re.compile(r"(try|catch|except|raise|throw|Error|Exception)", re.IGNORECASE), 5),
-    (re.compile(r"(error.?handl|fallback|retry|timeout)", re.IGNORECASE), 5),
-]
-
-# Category name -> base score for that category of fix (from CATEGORY_ORDER priority).
-_CATEGORY_SCORES: dict[str, int] = {
-    "Security": 8,
-    "Error Handling": 6,
-    "Tests": 6,
-    "A11y": 5,
-    "Code Quality": 3,
-    "Performance": 5,
-    "Polish": 2,
-}
-
-
-def _diff_line_score(diff_text: str) -> int:
-    """Scan diff added-lines for high-impact patterns. Return the max signal found."""
-    best = 0
-    for line in diff_text.splitlines():
-        if not line.startswith("+"):
-            continue
-        for pattern, score in _SECURITY_PATTERNS:
-            if pattern.search(line):
-                best = max(best, score)
-        for pattern, score in _ERROR_HANDLING_PATTERNS:
-            if pattern.search(line):
-                best = max(best, score)
-    return best
-
-
-def _has_test_files(files: list[str]) -> bool:
-    """Return True if any touched file looks like a test file."""
-    for f in files:
-        name = f.rsplit("/", 1)[-1].lower()
-        if name.startswith("test_") or name.endswith("_test.py") or name.endswith(".test.ts"):
-            return True
-        if name.endswith(".test.js") or name.endswith(".test.tsx") or name.endswith(".test.jsx"):
-            return True
-        if name.endswith(".spec.ts") or name.endswith(".spec.js"):
-            return True
-    return False
-
-
-def score_diff(
-    *,
-    worktree_dir: Path,
-    pre_head: str,
-    cycle_result: CycleResult | None,
-    files_touched: list[str],
-) -> DiffScore:
-    """Score a cycle's changes on a 1-10 scale based on production impact.
-
-    Scoring factors:
-    1. Category of fixes (from cycle_result) - higher-priority categories score higher
-    2. Diff content analysis - security/error patterns in added lines boost the score
-    3. Test bonus - writing tests adds +1
-    4. Category breadth bonus - multiple categories in one cycle adds +1
-    """
-    base = 1  # minimum score for any accepted cycle
-
-    # Factor 1: category scores from structured result
-    categories_seen: set[str] = set()
-    if cycle_result is not None:
-        for fix in cycle_result.get("fixes", []):
-            cat = fix.get("category", "")
-            if cat:
-                categories_seen.add(cat)
-                cat_score = _CATEGORY_SCORES.get(cat, 3)
-                base = max(base, cat_score)
-
-    # Factor 2: diff content analysis
-    try:
-        diff_text = git(worktree_dir, "diff", f"{pre_head}..HEAD")
-    except NightshiftError:
-        diff_text = ""
-    line_score = _diff_line_score(diff_text)
-    base = max(base, line_score)
-
-    # Factor 3: test bonus
-    test_bonus = _has_test_files(files_touched)
-    if test_bonus:
-        base = min(base + 1, 10)
-
-    # Factor 4: category breadth bonus
-    category_bonus = len(categories_seen) >= 2
-    if category_bonus:
-        base = min(base + 1, 10)
-
-    # Build reason string
-    parts: list[str] = []
-    if categories_seen:
-        parts.append(f"categories: {', '.join(sorted(categories_seen))}")
-    if line_score > 0:
-        parts.append(f"diff signal: {line_score}")
-    if test_bonus:
-        parts.append("test file included")
-    if category_bonus:
-        parts.append("multi-category breadth")
-    reason = "; ".join(parts) if parts else "no high-impact signals detected"
-
-    return DiffScore(
-        score=min(base, 10),
-        reason=reason,
-        category_bonus=category_bonus,
-        test_bonus=test_bonus,
-    )
